CalculationOfSquareRoot
